slow + boost functions

+++ TowerDefense/TowerDefense/EnemyEvade.cs
@@ -17,7 +17,7 @@ namespace TowerDefense
             set { isEvade = value; }
         }
         //Constructor
-        public EnemyEvade(string name, bool isEvade, float hp, int speed, int armor, int worth, Effect effect, string imagePath, PointF position, PointF endPosition, bool isClickable) : base(name,hp,speed,armor,worth,effect,imagePath,position,endPosition,isClickable)
+        public EnemyEvade(string name, bool isEvade, float hp, float speed, int armor, int worth, Effect effect, string imagePath, PointF position, PointF endPosition, bool isClickable) : base(name,hp,speed,armor,worth,effect,imagePath,position,endPosition,isClickable)
         {
             this.isEvade = isEvade;
         }
